Two unrelated fixes uncovered while wiring up stage-1 short training
1) utils/misc.init_wandb: skip wandb.login() for new wandb_v1_* personal-
   access tokens (>40 chars), which the SDK rejects as malformed. Set
   WANDB_API_KEY in env so wandb.init authenticates from netrc/env on its
   own. 40-char hex keys still go through wandb.login as before.

2) utils/data.apply_transforms_tensor_batch: cast new_pos_p to the
   destination dtype before the masked assign. Under bf16-mixed autocast
   the rotations/translations come back as bf16 while new_pos was cloned
   from the fp32 protein_pos, and the index_put errored. This path only
   runs in training_step's diagnostic block (rmsd_protein_pos logging)
   so it does not affect the loss graph; it just made every train_report
   iteration crash under bf16.

Verified: 10-step bf16-mixed stage-1 smoke now completes with
validation/loss=2.215 and a live wandb run.

+++ utils/data.py
@@ -591,7 +591,11 @@ def apply_transforms_tensor_batch(
         )
 
         # --- 把结果写回到总张量 ---
-        new_pos[atom_mask] = new_pos_p
+        # under bf16 autocast new_pos_p comes back as bf16 while new_pos was
+        # cloned from the fp32 input; cast to the destination dtype so the
+        # masked assign succeeds. This path runs in training_step's diagnostic
+        # block only, so it does not affect the loss graph.
+        new_pos[atom_mask] = new_pos_p.to(new_pos.dtype)
 
     return new_pos
